requirements.txt
fastapi
uvicorn
python-telegram-bot
yt-dlp
google-api-python-client
google-auth
google-auth-oauthlib
requests
